# docs/evidence/documentacion-etapa-2026-10-10/openapi-publico.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ashakids API",
    "description": "ASHAKids Backend REST API — Plataforma de telerehabilitación e intervención infantil.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/terapeutas": {
      "get": {
        "tags": [
          "Terapeutas"
        ],
        "summary": "Directorio",
        "operationId": "directorio_api_v1_terapeutas_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProfesionalSalida"
                  },
                  "title": "Response Directorio Api V1 Terapeutas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/terapeutas/{key}/disponibilidad": {
      "get": {
        "tags": [
          "Terapeutas"
        ],
        "summary": "Disponibilidad",
        "operationId": "disponibilidad_api_v1_terapeutas__key__disponibilidad_get",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisponibilidadDatos"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Terapeutas"
        ],
        "summary": "Publicar",
        "operationId": "publicar_api_v1_terapeutas__key__disponibilidad_put",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisponibilidadDatos"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisponibilidadDatos"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/terapeutas/{key}/turnos": {
      "get": {
        "tags": [
          "Terapeutas"
        ],
        "summary": "Turnos",
        "operationId": "turnos_api_v1_terapeutas__key__turnos_get",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Key"
            }
          },
          {
            "name": "fecha",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Fecha"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TurnoSalida"
                  },
                  "title": "Response Turnos Api V1 Terapeutas  Key  Turnos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/terapeutas/me": {
      "get": {
        "tags": [
          "Terapeutas"
        ],
        "summary": "Obtener perfil del terapeuta autenticado",
        "description": "Retorna la información del usuario autenticado y su perfil profesional. Requiere sesión activa y rol TERAPEUTA.",
        "operationId": "get_terapeuta_me_api_v1_terapeutas_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerapeutaProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/cuentas/terapeutas": {
      "post": {
        "tags": [
          "Administración"
        ],
        "summary": "Crear cuenta de Terapeuta",
        "description": "Crea de forma transaccional el Usuario con rol TERAPEUTA, su perfil en TERAPEUTAS y registra la auditoría. Requiere rol ADMIN.",
        "operationId": "post_crear_terapeuta_api_v1_admin_cuentas_terapeutas_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearTerapeutaRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperacionCuentaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BloqueoDatos": {
        "properties": {
          "inicio": {
            "type": "string",
            "format": "date-time",
            "title": "Inicio"
          },
          "fin": {
            "type": "string",
            "format": "date-time",
            "title": "Fin"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "inicio",
          "fin"
        ],
        "title": "BloqueoDatos"
      },
...
      "CrearTerapeutaRequest": {
        "properties": {
          "nombres": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Nombres"
          },
          "apellidos": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Apellidos"
          },
          "email": {
            "type": "string",
            "maxLength": 150,
            "minLength": 3,
            "title": "Email"
          },
          "password": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9]{8}$"
              },
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 12
              }
            ],
            "title": "Password"
          },
          "especialidad": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Especialidad"
          },
          "anios_experiencia": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 70.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Anios Experiencia"
          },
          "idiomas": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Idiomas"
          },
          "descripcion_profesional": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion Profesional"
          }
        },
        "type": "object",
        "required": [
          "nombres",
          "apellidos",
          "email",
          "password"
        ],
        "title": "CrearTerapeutaRequest",
        "description": "Payload para creación administrativa de Terapeuta."
      },
      "CuentaItemResponse": {
        "properties": {
          "id_usuario": {
            "type": "integer",
            "title": "Id Usuario"
          },
          "codigo_usuario": {
            "type": "string",
            "title": "Codigo Usuario"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "nombres": {
            "type": "string",
            "title": "Nombres"
          },
          "apellidos": {
            "type": "string",
            "title": "Apellidos"
          },
          "rol": {
            "type": "string",
            "title": "Rol"
          },
          "activo": {
            "type": "boolean",
            "title": "Activo"
          },
          "fecha_creacion": {
            "type": "string",
            "format": "date-time",
            "title": "Fecha Creacion"
          },
          "tutor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TutorData"
              },
              {
                "type": "null"
              }
            ]
          },
          "terapeuta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TerapeutaData"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id_usuario",
          "codigo_usuario",
          "email",
          "nombres",
          "apellidos",
          "rol",
          "activo",
          "fecha_creacion"
        ],
        "title": "CuentaItemResponse",
        "description": "Representación de una cuenta en el panel administrativo."
      },
...
      "DisponibilidadDatos": {
        "properties": {
          "turnos": {
            "items": {
              "$ref": "#/components/schemas/TurnoDatos"
            },
            "type": "array",
            "maxItems": 60,
            "title": "Turnos"
          },
          "bloqueos": {
            "items": {
              "$ref": "#/components/schemas/BloqueoDatos"
            },
            "type": "array",
            "maxItems": 60,
            "title": "Bloqueos"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "turnos"
        ],
        "title": "DisponibilidadDatos"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperacionCuentaResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "cuenta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CuentaItemResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "OperacionCuentaResponse",
        "description": "Respuesta tras realizar operaciones sobre cuentas."
      },
...
      "ProfesionalSalida": {
        "properties": {
          "id_terapeuta": {
            "type": "integer",
            "title": "Id Terapeuta"
          },
          "nombres": {
            "type": "string",
            "title": "Nombres"
          },
          "apellidos": {
            "type": "string",
            "title": "Apellidos"
          },
          "especialidad": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Especialidad"
          },
          "descripcion_profesional": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion Profesional"
          }
        },
        "type": "object",
        "required": [
          "id_terapeuta",
          "nombres",
          "apellidos",
          "especialidad",
          "descripcion_profesional"
        ],
        "title": "ProfesionalSalida"
      },
...
      "TerapeutaData": {
        "properties": {
          "id_terapeuta": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Terapeuta"
          },
          "especialidad": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Especialidad"
          },
          "anios_experiencia": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anios Experiencia"
          },
          "idiomas": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idiomas"
          },
          "descripcion_profesional": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion Profesional"
          }
        },
        "type": "object",
        "title": "TerapeutaData"
      },
      "TerapeutaProfileResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "perfil_terapeuta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TerapeutaData"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "user"
        ],
        "title": "TerapeutaProfileResponse",
        "description": "Respuesta para GET /api/v1/terapeutas/me."
      },
...
      "TurnoDatos": {
        "properties": {
          "dia": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 0.0,
            "title": "Dia"
          },
          "hora": {
            "type": "integer",
            "maximum": 17.0,
            "minimum": 8.0,
            "title": "Hora"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dia",
          "hora"
        ],
        "title": "TurnoDatos"
      },
      "TurnoSalida": {
        "properties": {
          "inicio": {
            "type": "string",
            "format": "date-time",
            "title": "Inicio"
          },
          "fin": {
            "type": "string",
            "format": "date-time",
            "title": "Fin"
          },
          "disponible": {
            "type": "boolean",
            "title": "Disponible"
          },
          "motivo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "inicio",
          "fin",
          "disponible"
        ],
        "title": "TurnoSalida"
      },
      "TutorData": {
        "properties": {
          "id_tutor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Tutor"
          },
          "parentesco": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parentesco"
          },
          "telefono": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          },
          "direccion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direccion"
          }
        },
        "type": "object",
        "title": "TutorData"
      },
      "UserResponse": {
        "properties": {
          "id_usuario": {
            "type": "integer",
            "title": "Id Usuario"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "nombres": {
            "type": "string",
            "title": "Nombres"
          },
          "apellidos": {
            "type": "string",
            "title": "Apellidos"
          },
          "codigo_usuario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codigo Usuario"
          },
          "rol": {
            "type": "string",
            "title": "Rol"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles",
            "default": []
          },
          "activo": {
            "type": "boolean",
            "title": "Activo",
            "default": true
          },
          "password_change_required": {
            "type": "boolean",
            "title": "Password Change Required",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id_usuario",
          "email",
          "nombres",
          "apellidos",
          "rol"
        ],
        "title": "UserResponse",
        "description": "Información segura del usuario autenticado (nunca expone password_hash)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}